feat(navigation): separate section component
