+++ src/modules/shared/components/Navigation/Navigation.tsx
@@ -1,108 +1,9 @@
-import { Button } from '@meemoo/react-components';
 import clsx from 'clsx';
-import { FC, useState } from 'react';
-
-import { Icon, IconLightNames } from '../Icon';
-import { Overlay } from '../Overlay';
+import { FC } from 'react';
 
 import styles from './Navigation.module.scss';
-import {
-	NavigationCenterProps,
-	NavigationFC,
-	NavigationProps,
-	NavigationSectionProps,
-} from './Navigation.types';
-import { NavigationDropdown } from './NavigationDropdown';
-import NavigationList from './NavigationList/NavigationList';
-
-const NavigationSection: FC<NavigationSectionProps> = ({
-	children,
-	items,
-	placement,
-	renderHamburger,
-	hamburgerProps,
-}) => {
-	const [isHamburgerMenuOpen, setIsHamburgerMenuOpen] = useState(false); // Note: Needed for overlay state. Dropdown state is saved in NavigationDropdown component
-
-	const renderHamburgerMenu = () => {
-		return (
-			<div className={styles['c-navigation__section--responsive-mobile']}>
-				{renderHamburger && (
-					<Overlay
-						visible={isHamburgerMenuOpen}
-						className={clsx(
-							styles['c-navigation__dropdown-overlay'],
-							styles['c-navigation__hamburger-overlay']
-						)}
-					/>
-				)}
-				<NavigationDropdown
-					id={'menu'}
-					isOpen={isHamburgerMenuOpen}
-					items={items ? items : []}
-					className={styles['c-navigation__hamburger']}
-					trigger={
-						<Button
-							label={
-								isHamburgerMenuOpen
-									? hamburgerProps?.hamburgerLabelOpen
-									: hamburgerProps?.hamburgerLabelClosed
-							}
-							variants="text"
-							className="u-color-white u-px-12 u-ml--12"
-							iconStart={
-								<Icon
-									className={clsx(
-										'u-fs-24',
-										`u-text-${placement}`,
-										!isHamburgerMenuOpen && 'u-color-teal'
-									)}
-									name={
-										isHamburgerMenuOpen
-											? hamburgerProps?.hamburgerIconOpen
-												? hamburgerProps?.hamburgerIconOpen
-												: 'times'
-											: hamburgerProps?.hamburgerIconClosed
-											? hamburgerProps?.hamburgerIconClosed
-											: 'grid-view'
-									}
-								/>
-							}
-						/>
-					}
-					lockScroll
-					flyoutClassName={clsx(
-						styles['c-navigation__dropdown-flyout'],
-						styles['c-navigation__hamburger']
-					)}
-					onOpen={() => setIsHamburgerMenuOpen(true)}
-					onClose={() => setIsHamburgerMenuOpen(false)}
-				/>
-			</div>
-		);
-	};
-
-	const renderDesktop = () => {
-		return renderHamburger ? (
-			<div className={styles['c-navigation__section--responsive-desktop']}>
-				{items && items.length ? <NavigationList items={items} /> : children}
-			</div>
-		) : (
-			<>{items && items.length ? <NavigationList items={items} /> : children}</>
-		);
-	};
-	return (
-		<div
-			className={clsx(
-				styles['c-navigation__section'],
-				styles[`c-navigation__section--${placement}`]
-			)}
-		>
-			{renderDesktop()}
-			{renderHamburger && renderHamburgerMenu()}
-		</div>
-	);
-};
+import { NavigationCenterProps, NavigationFC, NavigationProps } from './Navigation.types';
+import { NavigationSection } from './NavigationSection';
 
 const NavigationCenter: FC<NavigationCenterProps> = ({ children, title }) => (
 	<div className={styles['c-navigation__section']}>
