+++ src/modules/shared/components/Navigation/Navigation.types.ts
@@ -2,7 +2,7 @@ import { FC, ReactNode } from 'react';
 
 import { DefaultComponentProps } from '@shared/types';
 
-import { IconLightNames } from '../Icon';
+import { NavigationSectionProps } from './NavigationSection';
 
 export type NavigationFC<P = unknown> = FC<P> & {
 	Left: FC<NavigationSectionProps>;
@@ -18,24 +18,10 @@ export interface NavigationCenterProps {
 	title?: string;
 }
 
-export interface NavigationSectionProps {
-	items?: NavigationItem[];
-	placement: 'left' | 'right';
-	renderHamburger?: boolean;
-	hamburgerProps?: NavigationHamburgerProps;
-}
-
 export interface NavigationItem {
 	node: ReactNode;
 	id: string;
 	active?: boolean;
 	hasDivider?: boolean;
 	children?: NavigationItem[];
-}
-
-export interface NavigationHamburgerProps {
-	hamburgerLabelOpen: string;
-	hamburgerLabelClosed: string;
-	hamburgerIconOpen?: IconLightNames;
-	hamburgerIconClosed?: IconLightNames;
 }
