+++ src/modules/shared/components/Navigation/NavigationSection/index.ts
@@ -1,0 +1,2 @@
+export { default as NavigationSection } from './NavigationSection';
+export * from './NavigationSection.types';
