+++ src/modules/shared/components/Navigation/index.ts
@@ -1,2 +1,3 @@
 export { default as Navigation } from './Navigation';
 export * from './Navigation.types';
+export * from './NavigationSection';
